renames time pf2 decompostion files to adhere to new standard
